Always show AnalyzerName for AzureAISearch index profile (#18032)
